File restructure (#17)
* File restructure

* renamed file
