+++ pages/api/workout/[wid].tsx
