+++ pages/privacy.tsx
