--- src/components/lib/index.ts
@@ -1,3 +1,0 @@
-export function classNames(...classes: string[]) {
-    return classes.filter(Boolean).join(' ')
-  }
