+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "compilerOptions": {
     "target": "es6",
-    "baseUrl":"src",
+    "baseUrl":".",
     "lib": [
       "dom",
       "dom.iterable",
